Fix build temporarily

+++ src/arduino/display.cpp
@@ -46,7 +46,7 @@ void startScreen()
     }
 }
 
-void displayEveryRespiratoryCycle(uint16_t peakPressure, uint16_t plateauPressure, uint16_t peep)
+void displayEveryRespiratoryCycle(int peakPressure, int plateauPressure, int peep)
 {
     screen.setCursor(0, 0);
 
@@ -77,10 +77,7 @@ void displayEveryRespiratoryCycle(uint16_t peakPressure, uint16_t plateauPressur
     }
 }
 
-void displayDuringCycle(uint16_t cyclesPerMinute,
-                        uint16_t maxPlateauPressure,
-                        uint16_t peep,
-                        uint16_t currentPressure)
+void displayDuringCycle(int cyclesPerMinute, int maxPlateauPressure, int peep, int currentPressure)
 {
     screen.setCursor(0, 1);
